Lightning nd Thunder

+++ userbot/plugins/alive.py
@@ -1,17 +1,32 @@
 
-
-"""Check if 𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤 alive"""
-# CREDITS: @WhySooSerious, @Sur_vivor
+# For @borgHelp
+"""Check if your userbot is working."""
 import time
+from datetime import datetime
+from io import BytesIO
 
-from userbot.utils import admin_cmd, sudo_cmd
+import requests
+from PIL import Image
 
-from userbot import ALIVE_NAME, Lastupdate
-from userbot.Config import Var
-from userbot.plugins import currentversion
+from userbot import ALIVE_NAME, CMD_HELP, telever
+from userbot.__init__ import StartTime
+from userbot.Config import Config, Var
+
+# ======CONSTANTS=========#
+CUSTOM_ALIVE = (
+    Var.CUSTOM_ALIVE
+    if Var.CUSTOM_ALIVE
+    else "Hey! I'm alive. All systems online and functioning normally!"
+)
+ALV_PIC = Var.ALIVE_PIC if Var.ALIVE_PIC else None
+telemoji = Var.CUSTOM_ALIVE_EMOJI if Var.CUSTOM_ALIVE_EMOJI else "**✵**"
+if Config.SUDO_USERS:
+    sudo = "Enabled"
+else:
+    sudo = "Disabled"
+# ======CONSTANTS=========#
 
 
-# Functions
 def get_readable_time(seconds: int) -> str:
     count = 0
     ping_time = ""
@@ -40,29 +55,64 @@ def get_readable_time(seconds: int) -> str:
     return ping_time
 
 
-uptime = get_readable_time((time.time() - Lastupdate))
-DEFAULTUSER = str(ALIVE_NAME) if ALIVE_NAME else "Unknown"
-PM_IMG = Var.ALIVE_IMAGE
-pm_caption = "➥ **𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤  IS:** `ONLINE`\n\n"
-pm_caption += "➥ **SYSTEMS STATS**\n"
-pm_caption += "➥ **Telethon Version:** `1.15.0` \n"
-pm_caption += "➥ **Python:** `3.9.0` \n"
-pm_caption += f"➥ **Uptime** : `{uptime}` \n"
-pm_caption += "➥ **Database Status:**  `Functional`\n"
-pm_caption += "➥ **Current Branch** : `master`\n"
-pm_caption += f"➥ **Version** : `{currentversion}`\n"
-pm_caption += f"➥ **My Boss** : {DEFAULTUSER} \n"
-pm_caption += "➥ **Heroku Database** : `AWS - Working Properly`\n\n"
-pm_caption += "➥ **License** : [GNU General Public License v3.0](github.com/Anmol-dot283/black/blob/master/LICENSE)\n"
-pm_caption += "➥ **Copyright** : By [Krish@Github](GitHub.com/Anmol-dot283)\n"
-pm_caption += "➥ **Check Stats By Doing** `.stat`. \n\n"
-pm_caption += "[🇮🇳 Deploy 𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤 Userbot 🇮🇳](https://telegra.ph/file/c3683ea84426c1046faef.mp4)"
+DEFAULTUSER = str(ALIVE_NAME) if ALIVE_NAME else "@blacklightningot"
 
 
-@borg.on(admin_cmd(pattern=r"alive"))
-@borg.on(sudo_cmd(pattern=r"alive", allow_sudo=True))
-async def black(alive):
-    await alive.get_chat()
+@borg.on(admin_cmd(outgoing=True, pattern="alive"))
+@borg.on(sudo_cmd(outgoing=True, pattern="alive", allow_sudo=True))
+async def amireallyalive(alive):
+    start = datetime.now()
+    myid = bot.uid
     """ For .alive command, check if the bot is running.  """
-    await borg.send_file(alive.chat_id, PM_IMG, caption=pm_caption)
-    await alive.delete()
+    end = datetime.now()
+    (end - start).microseconds / 1000
+    uptime = get_readable_time((time.time() - StartTime))
+    if ALV_PIC:
+        tele = f"**Welcome To 𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤 **\n\n"
+        tele += f"`{CUSTOM_ALIVE}`\n\n"
+        tele += (
+            f"{telemoji} **Telethon version**: `1.17`\n{telemoji} **Python**: `3.9`\n"
+        )
+        tele += f"{telemoji} **𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤 Version**: `{telever}`\n"
+        tele += f"{telemoji} **More Info**: @blacklightning\n"
+        tele += f"{telemoji} **Sudo** : `{sudo}`\n"
+        tele += f"{telemoji} **Thunder Uptime**: `{uptime}`\n"
+        tele += f"{telemoji} **Database Status**: `All OK 👌!`\n"
+        tele += (
+            f"{telemoji} **My pro owner** : [{DEFAULTUSER}](tg://user?id={myid})\n\n"
+        )
+        tele += "    [✨ GitHub Repository ✨](https://github.com/Anmol-dot283/Black-Lightning)"
+        await alive.get_chat()
+        await alive.delete()
+        """ For .alive command, check if the bot is running.  """
+        await borg.send_file(alive.chat_id, ALV_PIC, caption=tele, link_preview=False)
+        await alive.delete()
+        return
+    req = requests.get("https://telegra.ph/file/c3683ea84426c1046faef.mp4")
+    req.raise_for_status()
+    file = BytesIO(req.content)
+    file.seek(0)
+    img = Image.open(file)
+    with BytesIO() as sticker:
+        img.save(sticker, "webp")
+        sticker.name = "sticker.webp"
+        sticker.seek(0)
+        await borg.send_message(
+            alive.chat_id,
+            f"**Welcome To 𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤  **\n\n"
+            f"`{CUSTOM_ALIVE}`\n\n"
+            f"{telemoji} **Telethon version**: `1.17`\n{telemoji} **Python**: `3.9`\n"
+            f"{telemoji} **𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤  Version**: `{telever}`\n"
+            f"{telemoji} **More Info**:` @borgSupport\n"
+            f"{telemoji} **Sudo** : `{sudo}`\n"
+            f"{telemoji} **𝔅𝔩𝔞𝔠𝔨 𝔏𝔦𝔤𝔥𝔱𝔫𝔦𝔫𝔤  Uptime**: `{uptime}`\n"
+            f"{telemoji} **Database Status**: `All OK 👌!`\n"
+            f"{telemoji} **My pro owner** : [{DEFAULTUSER}](tg://user?id={myid})\n\n"
+            "    [✨ GitHub Repository ✨](https://github.com/Anmol-dot283/Black-Lightning)",
+            link_preview=False,
+        )
+        await borg.send_file(alive.chat_id, file=sticker)
+        await alive.delete()
+
+
+CMD_HELP.update({"alive": "➟ `.alive`\nUse - Check if your bot is working."})
